Check if directory exists before deleting temp folder
Added a conditional check to verify the existence of the temporary folder before attempting deletion. This change prevents potential exceptions that may occur if the folder does not exist, improving the application's stability.

+++ SimpleLauncher/ExtractCompressedFile.cs
@@ -126,7 +126,10 @@ namespace SimpleLauncher
             try
             {
                 // Delete temp folder
-                Directory.Delete(_tempFolder, true);
+                if (Directory.Exists(_tempFolder))
+                {
+                    Directory.Delete(_tempFolder, true);
+                }
             }
             catch (Exception ex)
             {
